refactor: make user's viewModel package and seperate
