+++ app/src/main/java/com/fruitable/Fruitable/app/presentation/view/user/LogInScreen.kt
@@ -32,7 +32,7 @@ import com.fruitable.Fruitable.app.presentation.component.FruitableButton
 import com.fruitable.Fruitable.app.presentation.component._feature.TextFieldBox
 import com.fruitable.Fruitable.app.presentation.event.LogInEvent
 import com.fruitable.Fruitable.app.presentation.navigation.Screen
-import com.fruitable.Fruitable.app.presentation.viewmodel.LogInViewModel
+import com.fruitable.Fruitable.app.presentation.viewmodel.user.LogInViewModel
 import com.fruitable.Fruitable.ui.theme.MainGray6
 import com.fruitable.Fruitable.ui.theme.MainGreen1
 import com.fruitable.Fruitable.ui.theme.TextStyles
